Moved the code over to broadcast.cpp
